move helpers and constants to support so cypress doesn't try to run tests in those files in CI
