+++ e2e/cypress/integration/api/iam/02_policies_api.spec.ts
@@ -1,4 +1,4 @@
-import { describeIfIAMV2p1 } from '../../constants';
+import { describeIfIAMV2p1 } from '../../../support/constants';
 
 describeIfIAMV2p1('policies API', () => {
   const defaultAdminReq = {
