+++ e2e/cypress/integration/api/iam/03_iam_integration.spec.ts
@@ -1,4 +1,4 @@
-import { describeIfIAMV2p1 } from '../../constants';
+import { describeIfIAMV2p1 } from '../../../support/constants';
 
 describeIfIAMV2p1('assigning projects', () => {
   let twoAllowedProjectsTok = '';
