+++ e2e/cypress/integration/api/iam/04_projects_api_client_runs_ingestion.spec.ts
@@ -1,5 +1,5 @@
-import { describeIfIAMV2p1 } from '../../constants';
-import { uuidv4 } from '../../helpers';
+import { describeIfIAMV2p1 } from '../../../support/constants';
+import { uuidv4 } from '../../../support/helpers';
 
 describeIfIAMV2p1('Client Runs Ingestion project tagging', () => {
   const cypressPrefix = 'test-projects-api-client-runs';
