+++ e2e/cypress/integration/api/iam/05_projects_api_compliance_ingestion.spec.ts
@@ -1,5 +1,5 @@
-import { describeIfIAMV2p1 } from '../../constants';
-import { uuidv4 } from '../../helpers';
+import { describeIfIAMV2p1 } from '../../../support/constants';
+import { uuidv4 } from '../../../support/helpers';
 
 describeIfIAMV2p1('Compliance Ingestion project tagging', () => {
   const cypressPrefix = 'test-compliance-ingestion';
