+++ e2e/cypress/integration/api/iam/06_projects_api_compliance_update.spec.ts
@@ -1,5 +1,5 @@
-import { describeIfIAMV2p1 } from '../../constants';
-import { uuidv4 } from '../../helpers';
+import { describeIfIAMV2p1 } from '../../../support/constants';
+import { uuidv4 } from '../../../support/helpers';
 
 describeIfIAMV2p1('Compliance report project update tagging', () => {
   const cypressPrefix = 'test-compliance-update';
