+++ e2e/cypress/integration/ui/common/02_global_filter.spec.ts
@@ -1,4 +1,4 @@
-import { describeIfIAMV2p1 } from '../../constants';
+import { describeIfIAMV2p1 } from '../../../support/constants';
 
 interface CreateProject {
   id: string;
