+++ e2e/cypress/integration/ui/settings/01_user_mgmt.spec.ts
@@ -1,4 +1,4 @@
-import { itFlaky } from '../../constants';
+import { itFlaky } from '../../../support/constants';
 
 describe('user management', () => {
   before(() => {
