+++ e2e/cypress/integration/ui/settings/03_team_add_users.spec.ts
@@ -1,4 +1,4 @@
-import { isV1, itFlaky } from '../../constants';
+import { isV1, itFlaky } from '../../../support/constants';
 
 describe('team add users', () => {
   const now = Cypress.moment().format('MMDDYYhhmm');
