+++ e2e/cypress/integration/ui/settings/04_team_details.spec.ts
@@ -1,4 +1,4 @@
-import { describeIfIAMV2p1, isV1, itFlaky } from '../../constants';
+import { describeIfIAMV2p1, isV1, itFlaky } from '../../../support/constants';
 
 describe('team details', () => {
   let adminIdToken = '';
