+++ e2e/cypress/integration/ui/settings/06_token_mgmt.spec.ts
@@ -1,4 +1,4 @@
-import { iamVersion } from '../../constants';
+import { iamVersion } from '../../../support/constants';
 
 describe('token management', () => {
   const typeDelay = 50;
