+++ e2e/cypress/support/index.ts
@@ -15,7 +15,7 @@
 
 // Import commands.js using ES2015 syntax:
 import './commands';
-import '../integration/constants';
+import './constants';
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
